refactor: move export modal component to components directory
